Decowpled the robot controller from the communication but adding an EventDispatcher interface to increase maintainability and flexibility.
